[processing] use the expiry weekday and week num rather than epoch timestamp
We probably can't glean much at all from the raw timestamp as we could from
the fact that this expiry was on a wednesday or that it was the third
friday of the month.

Additionally, make sure that the no normalization is applied to these
columns.

Finally, move all DataFrame processing into the one function. The other
constituent functions weren't being used individually.